- Updated SnakebiteIncidentService to handle rescuer shift assignments more accurately, including overnight shifts.
- Refactored mission creation logic to simplify request updates.
- Introduced VenomTypeService for managing venom types, including creation, retrieval, updating, and deletion.
- Added ISnakebiteIncidentPaymentService interface for handling payment links and transactions related to snakebite incidents.
- Improved TreatmentFacilityService to ensure unique treatment facility IDs and included active status in responses.
- Updated ILocationIqService and IShiftService interfaces to support new pricing calculations and bulk shift assignments.
- Added unit tests for ShiftAssignmentExtensions to validate duty status logic.
- Documented workshift and shift assignment APIs for better clarity and usage guidance.

+++ SnakeAid.Service/Services/PayOs/Models/PayOsPaymentLinkResult.cs
@@ -6,8 +6,8 @@ public class PayOsPaymentLinkResult
     public string PaymentLinkId { get; set; } = string.Empty;
     public string CheckoutUrl { get; set; } = string.Empty;
     public decimal Amount { get; set; }
-    public int AmountPaid { get; set; }
-    public int AmountRemaining { get; set; }
+    public decimal AmountPaid { get; set; }
+    public decimal AmountRemaining { get; set; }
     public string Status { get; set; } = string.Empty;
     public string Currency { get; set; } = "VND";
     public bool Success { get; set; }
